Fix type

+++ mobile-go/main.go
@@ -71,10 +71,10 @@ func (saver AlarmSaver) saveAlarm(alarmInfo map[string]interface{}) {
 	alarmInfo["proofList"] = mediaList
 
 	dateStr := time.Now().Format("20060102150405")
-	alarmType := alarmInfo["alarmType"].(int)
+	alarmType := alarmInfo["alarmType"].(float64)
 	devidno := alarmInfo["devidno"]
 
-	fileName := fmt.Sprintf("%s_%d_%s.json", devidno, alarmType, dateStr)
+	fileName := fmt.Sprintf("%s_%d_%s.json", devidno, int(alarmType), dateStr)
 
 	byteInfo, err := json.MarshalIndent(alarmInfo, "", "  ")
 	failOnError(err, "Marshal error")
